% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decodePic}
\alias{decodePic}
\title{decodePic}
\usage{
decodePic(data)
}
\arguments{
\item{data}{pointer to array of bytes to be decoded (need memorycont. repr.)}
}
\value{
the number of decoded doubles
}
\description{
Decodes data encoded by encodePic
  
  result vector guaranteed to be shorter of equal to |data| * 2
}
\details{
Note that this method may throw a const char* if it deems the input data to be corrupt, i.e.
that the last encoded int does not use the last byte in the data. In addition the last encoded 
  int need to use either the last halfbyte, or the second last followed by a 0x0 halfbyte.
}
\seealso{
[\code{\link{encodePic}}]
}
