% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{encodePic}
\alias{encodePic}
\title{encodePic}
\usage{
encodePic(data)
}
\arguments{
\item{data}{pointer to array of double to be encoded (need memorycont. repr.)}
}
\value{
the number of encoded bytes
}
\description{
Encodes ion counts by simply rounding to the nearest 4 byte integer, 
and compressing each integer with encodeInt.
}
\details{
The handleable range is therefore 0 -> 4294967294.
The resulting binary is maximally dataSize * 5 bytes, but much less if the 
  data is close to 0 on average.
}
\seealso{
[\code{\link{decodePic}}]
}
