% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimalLinearFixedPointMass}
\alias{optimalLinearFixedPointMass}
\title{optimalLinearFixedPointMass}
\usage{
optimalLinearFixedPointMass(data, mass_acc)
}
\arguments{
\item{data}{pointer to array of double to be encoded (need memorycont. repr.)}

\item{mass_acc}{desired m/z accuracy in Th}
}
\value{
the linear fixed point that satisfies the accuracy requirement (or -1 in case of failure).
}
\description{
Compute the optimal linear fixed point with a desired m/z accuracy.
}
\note{
If the desired accuracy cannot be reached without overflowing 64
bit integers, then a negative value is returned. You need to check for
  this and in that case abandon numpress or use optimalLinearFixedPoint
  which returns the largest safe value.
}
