% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimalSlofFixedPoint}
\alias{optimalSlofFixedPoint}
\title{optimalSlofFixedPoint}
\usage{
optimalSlofFixedPoint(data)
}
\arguments{
\item{data}{pointer to array of double to be encoded (need memorycont. repr.)}
}
\value{
the slof fixed point safe to use
}
\description{
Compute the maximal natural logarithm fixed point that prevents integer overflow.
}
